Fix workspace creation code
